server/requirements.txt
flask
flask-cors
psycopg2
load_dotenv